refactor: 이미지 응답 dto 패키지 변경
