+++ src/main/java/org/chzz/market/domain/auctionv2/repository/AuctionV2QueryRepository.java
@@ -46,8 +46,8 @@ import org.chzz.market.domain.auctionv2.dto.response.WonAuctionResponse;
 import org.chzz.market.domain.auctionv2.entity.AuctionStatus;
 import org.chzz.market.domain.auctionv2.entity.Category;
 import org.chzz.market.domain.bid.entity.QBid;
-import org.chzz.market.domain.image.dto.ImageResponse;
-import org.chzz.market.domain.image.dto.QImageResponse;
+import org.chzz.market.domain.imagev2.dto.response.ImageResponse;
+import org.chzz.market.domain.imagev2.dto.response.QImageResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.support.PageableExecutionUtils;
